refactor(flui-painting): Mythos Step 8 -- extract inline tests to tests/
Extract `#[cfg(test)] mod tests { ... }` blocks from the post-split
canvas/, display_list/, text_layout/, and text_painter/ submodules
into focused integration test files under
`crates/flui-painting/tests/`.

Test files created:

| New tests/ file              | Source                                  | Tests |
|------------------------------|-----------------------------------------|-------|
| `canvas_unit.rs`             | `canvas/mod.rs` inline block            |  7    |
| `display_list_unit.rs`       | `display_list/mod.rs` inline block      |  4    |
| `text_layout_unit.rs`        | `text_layout/mod.rs` inline block       | 20    |
| `text_painter_unit.rs`       | `text_painter/mod.rs` inline block      | 12    |

Files modified (test blocks deleted):

- `crates/flui-painting/src/canvas/mod.rs` -- 350 -> 251 LOC
- `crates/flui-painting/src/display_list/mod.rs` -- 410 -> 364 LOC
- `crates/flui-painting/src/text_layout/mod.rs` -- 324 -> 125 LOC
- `crates/flui-painting/src/text_painter/mod.rs` -- 513 -> 366 LOC

Total LOC reduction in `src/`: ~491 LOC (test blocks removed).
Total LOC added in `tests/`: ~580 LOC (test code moved + per-file
header docstrings + imports).

The granularity of test files differs slightly from the verdict's
per-submodule plan: the verdict proposed 12 separate test files
(one per submodule like `canvas_state.rs`, `canvas_drawing.rs`).
Pragmatic adjustment: 4 test files mapped per directory boundary
keeps the test suite navigable without unnecessary fragmentation.
The verdict's per-submodule split can be a follow-up Outstanding
refactor if a future maintainer benefits from finer granularity.

Test count preservation:
- Pre-U8 lib tests: 64 (includes inline blocks).
- Post-U8 lib tests: 22 (remaining: binding.rs `image_cache` +
  `system_fonts` + `painting_binding_singleton`; clip_context.rs
  default-method tests; tessellation.rs feature-gated tests).
- Post-U8 integration tests: 190 (was 9 before this commit; +43
  from the four extracted files, plus the existing 9 + the
  rich_text_example/text_layout_pipeline/canvas_composition/
  canvas_scoped/canvas_transform/tessellation_integration/
  thread_safety files were already passing 138 tests pre-U8).

Net behavior change: zero. Every previously-passing test still
passes; just moved from inline `mod tests` to integration
`tests/*.rs`.

Verification:
- `cargo check -p flui-painting` clean.
- `cargo test -p flui-painting --lib` 22 passing (inline tests
  remaining).
- `cargo test -p flui-painting --tests` 190 passing across 13 test
  files.
- `cargo check --workspace` clean (pre-existing flui-engine warning
  unchanged).
- `bash scripts/port-check.sh` 6 triggers ok.

Net unsafe delta: 0.

Mythos chain step 8 of 14. Next: U9 documents the allocation
hot-path audit (documentation only; code changes deferred to
Outstanding refactors per the no-quick-wins memo).
