+++ crates/flui-painting/src/display_list/mod.rs
@@ -362,49 +362,3 @@ impl IndexMut<usize> for DisplayList {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use flui_types::{geometry::px, styling::Color};
-
-    use super::*;
-
-    #[test]
-    fn test_display_list_creation() {
-        let display_list = DisplayList::new();
-        assert!(display_list.is_empty());
-        assert_eq!(display_list.len(), 0);
-        assert_eq!(display_list.bounds(), Rect::ZERO);
-    }
-
-    #[test]
-    fn test_display_list_push() {
-        let mut display_list = DisplayList::new();
-        let rect = Rect::from_ltrb(px(0.0), px(0.0), px(100.0), px(100.0));
-        let paint = Paint::fill(Color::RED);
-
-        display_list.push(DrawCommand::DrawRect {
-            rect,
-            paint,
-            transform: Matrix4::identity(),
-        });
-
-        assert_eq!(display_list.len(), 1);
-        assert_eq!(display_list.bounds(), rect);
-    }
-
-    #[test]
-    fn test_display_list_clear() {
-        let mut display_list = DisplayList::new();
-        display_list.push(DrawCommand::DrawRect {
-            rect: Rect::from_ltrb(px(0.0), px(0.0), px(100.0), px(100.0)),
-            paint: Paint::default(),
-            transform: Matrix4::identity(),
-        });
-
-        assert!(!display_list.is_empty());
-
-        display_list.clear();
-        assert!(display_list.is_empty());
-        assert_eq!(display_list.bounds(), Rect::ZERO);
-    }
-}
